+++ src/main/java/ceu/biolab/cmm/msSearch/dto/RTSearchResponseDTO.java
@@ -1,0 +1,33 @@
+package ceu.biolab.cmm.msSearch.dto;
+
+import ceu.biolab.cmm.shared.domain.msFeature.AnnotatedFeature;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class RTSearchResponseDTO {
+    private List<AnnotatedFeature> msFeatures;
+
+    public RTSearchResponseDTO() {
+        this.msFeatures = new ArrayList<>();
+    }
+
+    public void addImFeature(AnnotatedFeature imFeature) {
+        if (imFeature != null) {
+            this.msFeatures.add(imFeature);
+        }
+    }
+
+    public List<AnnotatedFeature> getMSFeatures() {
+        return msFeatures;
+    }
+
+    public void setmsFeatures(List<AnnotatedFeature> msFeatures) {
+        this.msFeatures = msFeatures;
+    }
+
+    @Override
+    public String toString() {
+        return "RTSearchResponse [Features=" + msFeatures + "]";
+    }
+}
