+++ src/main/java/ceu/biolab/cmm/msSearch/model/msFeature/LCMSFeature.java
@@ -1,0 +1,7 @@
+package ceu.biolab.cmm.msSearch.model.msFeature;
+
+public interface LCMSFeature {
+
+    public double getRT();
+
+}
