+++ src/main/java/ceu/biolab/cmm/msSearch/model/msFeature/LCMSMSFeature.java
@@ -1,0 +1,16 @@
+package ceu.biolab.cmm.msSearch.model.msFeature;
+
+import java.util.List;
+
+public class LCMSMSFeature implements LCMSFeature, MSMSFeature{
+
+    @Override
+    public double getRT() {
+        return 0;
+    }
+
+    @Override
+    public List<MSPeak> getProductIons() {
+        return null;
+    }
+}
