+++ src/main/java/ceu/biolab/cmm/msSearch/model/msFeature/MSMSFeature.java
@@ -1,0 +1,9 @@
+package ceu.biolab.cmm.msSearch.model.msFeature;
+
+import java.util.List;
+
+public interface MSMSFeature {
+
+    public List<MSPeak> getProductIons();
+
+}
