+++ src/main/java/ceu/biolab/cmm/msSearch/model/msFeature/MSMSFeatureImpl.java
@@ -1,0 +1,13 @@
+package ceu.biolab.cmm.msSearch.model.msFeature;
+
+import java.util.List;
+
+public class MSMSFeatureImpl implements MSMSFeature{
+
+
+    @Override
+    public List<MSPeak> getProductIons() {
+        return null;
+    }
+
+}
