+++ src/main/java/ceu/biolab/cmm/rtSearch/domain/compositeSpectra.java
@@ -1,0 +1,4 @@
+package ceu.biolab.cmm.rtSearch.domain;
+
+public class compositeSpectra {
+}
