--- src/main/java/ceu/biolab/cmm/rtSearch/model/msFeature/GCMSFeature.java
@@ -1,16 +1,0 @@
-package ceu.biolab.cmm.rtSearch.model.msFeature;
-
-import ceu.biolab.cmm.shared.domain.msFeature.MSFeature;
-
-public class GCMSFeature extends MSFeature {
-    private double rt;
-    private double ri;
-
-    public GCMSFeature(double mz, double intensity, double rt, double ri) {
-        super(mz, intensity);
-        this.rt = rt;
-        this.ri = ri;
-    }
-
-
-}
