--- src/main/java/ceu/biolab/cmm/rtSearch/model/msFeature/LCMSFeature.java
@@ -1,7 +1,0 @@
-package ceu.biolab.cmm.rtSearch.model.msFeature;
-
-public interface LCMSFeature {
-
-    public double getRT();
-
-}
