--- src/main/java/ceu/biolab/cmm/rtSearch/model/msFeature/MSMSFeature.java
@@ -1,9 +1,0 @@
-package ceu.biolab.cmm.rtSearch.model.msFeature;
-
-import java.util.List;
-
-public interface MSMSFeature {
-
-    public List<MSPeak> getProductIons();
-
-}
