--- src/main/java/ceu/biolab/cmm/rtSearch/model/msFeature/MSMSFeatureImpl.java
@@ -1,13 +1,0 @@
-package ceu.biolab.cmm.rtSearch.model.msFeature;
-
-import java.util.List;
-
-public class MSMSFeatureImpl implements MSMSFeature{
-
-
-    @Override
-    public List<MSPeak> getProductIons() {
-        return null;
-    }
-
-}
