+++ src/main/java/ceu/biolab/cmm/shared/domain/compound/Compound.java
@@ -1,7 +1,7 @@
 package ceu.biolab.cmm.shared.domain.compound;
 
 import ceu.biolab.cmm.shared.domain.FormulaType;
-import ceu.biolab.cmm.rtSearch.model.compound.LipidMapsClassification;
+import ceu.biolab.cmm.msSearch.model.compound.LipidMapsClassification;
 import lombok.Data;
 import lombok.experimental.SuperBuilder;
